+++ NEWS.md
@@ -3,6 +3,8 @@
 - Allow to reconstruct the original text + allow to add a start/end field in as.data.frame (useful but undocumented feature). Set up mainly to be used with the crfsuite R package
 - Added txt_tagsequence
 - Added 1 general function called udpipe which does annotation of data in TIF format.
+- Add option in udpipe_download_model to download the model only it does not exist on disk
+- Last loaded model is put into an environment
 
 # CHANGES IN udpipe VERSION 0.6.1